SettingsView : the reading of settings was refactored into an abstract settings_reader
The idea is that settings will be saved in the file as build in types (+strings) and the settings_reader will read settings into values of this types. This values will be then converted by the types::parse method to the types which are used for the settings representation. The types::parse method is trivial when the build in type is the same as the representation type or is convertible (as is the case for the enum class). For special cases a specialization of struct types template needs to be provided.