SettingsView : observer registration added to the settings_provider class
Observer will be notified always when settings_view::get_view method is called. The various observers are stored in a callback_container class which encapsulates registration, unregistration and calling of the callbacks. All these three methods are thread safe. Thread safe access to an instance is encapsulated by the monitor class.
The callback_container class can be used with the observer pattern using the observer_adapter class. This is more effective then using std::bind.